conslidated fixed noise functions into mx_noise.js
